requirements.txt
anyjson==0.3.3
Django==1.8.5
django-cors-headers==1.1.0
django-filter==0.11.0
djangorestframework==3.2.4
djangorestframework-jsonapi==2.0.1
djangorestframework-jwt==1.7.2
drf-nested-routers
djangorestframework-xml
PyYAML===3.11
django-cleanup===0.4.2
requests===2.11.1